pyproject.toml
[project]
name = "wallet"
version = "0.1.0"
description = ""
authors = [
    {name = "RocketMenace",email = "grizzlygit777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dishka (>=1.7.2,<2.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "uvloop (>=0.22.1,<0.23.0)",
    "httptools (>=0.7.1,<0.8.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "fastapi (>=0.119.1)",
    "uvicorn (>=0.38.0,<0.39.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.1,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]
